backend/pyproject.toml
[tool.poetry]
name = "first"
version = "0.1.0"
description = ""
authors = ["Yagunov Denis <spb.jagunov.denis@gmail.com>"]

[tool.poetry.dependencies]
python = "~=3.9"
asgiref = "^3.4.1"
certifi = "^2021.5.30"
cffi = "^1.14.6"
charset-normalizer = "^2.0.4"
coreapi = "^2.3.3"
coreschema = "^0.0.4"
coverage = "^5.5"
cryptography = "^3.4.7"
defusedxml = "^0.7.1"
Django = "^3.2.6"
django-cors-headers = "^3.8.0"
django-filter = "^2.4.0"
django-templated-mail = "^1.1.1"
djangorestframework = "^3.12.4"
djangorestframework-simplejwt = "^4.7.2"
djoser = "^2.1.0"
drf-extra-fields = "^3.1.1"
flake8 = "^3.9.2"
gunicorn = "^20.0.4"
idna = "^3.2"
isort = "^5.9.3"
itypes = "^1.2.0"
Jinja2 = "^3.0.1"
MarkupSafe = "^2.0.1"
mccabe = "^0.6.1"
oauthlib = "^3.1.1"
Pillow = "^8.3.1"
psycopg2-binary = "^2.9.1"
pycodestyle = "^2.7.0"
pycparser = "^2.20"
pyflakes = "^2.3.1"
PyJWT = "^2.1.0"
python-dotenv = "^0.19.0"
python3-openid = "^3.2.0"
pytz = "^2021.1"
requests = "^2.26.0"
requests-oauthlib = "^1.3.0"
six = "^1.16.0"
social-auth-app-django = "^4.0.0"
social-auth-core = "^4.1.0"
sqlparse = "^0.4.1"
uritemplate = "^3.0.1"
urllib3 = "^1.26.6"
drf-writable-nested = "^0.7.0"
django-phonenumber-field = "^7.3.0"
django-phonenumbers = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
